Upgrade to 1.21.8

+++ common/src/main/java/com/xtracr/realcamera/util/MathUtil.java
@@ -28,6 +28,6 @@ public class MathUtil {
         Vector4f vector4f = new Vector4f((float) vec3.x(), (float) vec3.y(), (float) vec3.z(), 1.0f);
         for (Matrix4f matrix4f : projectionMatrices) vector4f.mul(matrix4f);
         if (vector4f.w() == 0.0) return Vec3.ZERO;
-        return new Vec3(vector4f.x(), vector4f.y(), 0).scale(1 / (double) vector4f.w());
+        return new Vec3(vector4f.x(), vector4f.y(), 0).scale(1 / vector4f.w());
     }
 }
